style: rename window pixmap class to 'buffer'
The render::window_pixmap class wraps buffer-like objects on X11 and Wayland.
Because of its X11 legacy the class was originally named WindowPixmap, later
the name was snake-cased. This is a complicated name and it's difficult to
differentiate the class from the window class. Instead simply rename it to
'buffer', what simplifies the name while expressing meaning and usage of the
class just fine.