feat(effect): remove logging category declaration
The KWINEFFECTS logging category is not used by any of our internal effects.
It is an unnecessary declaration on a generic level. We remove it and accept
the risk of breaking external effects that depended on it. These can be easily
fixed by declaring their own logging categories instead.

BREAKING CHANGE: KWINEFFECTS logging category declaration is removed