ci: switch to stable image on beta release
Switch the CI to use the stable image on the new stable branch when preparing
the beta release. This is done in a separate commit on the stable branch after
it has been branched off from master.

In our CI pipeline we can now simply rely on the IMAGE_VERSION what allows to
create merge requests against stable branches.

Note that we still check on the CI_COMMIT_TAG as we tag the branching off
commit without the commit afterwards to change the variable but we still want
to run the pipeline on the tag in the release context with KF5 built from the
last stable release.