refactor: remove client list from space
We want a single list of all windows for simplicity. Remove the container with
windows with control and instead loop over the list of all windows checking for
control on each element. This is acceptable as there exist usually only a
handful of windows in the space at any point in time.