refactor: inhibit idle per window in free functions and without wrapper class
Create Wayland globals in workspace. Delete idle inhibition wrapper class. This
is possible because the Wrapland idle class already tracks the number of
inhibitions.

Wayland windows can inhibit idle when they are visible, but not otherwise.
We manage these check with free functions and a new boolean attribute set on
Wayland windows. The caller ensures the windows checked are of the correct type
and have control.