refactor(input): handle shortcuts in platform
Move the ownership from redirect to platform. This allows us to delay the
creation of redirect later on what otherwise is problematic together with the
KGlobalAccel Qt plugin in our tests. It also seems to be the better place in
general as the redirect class should only be about redirecting input and is
basically useless on X11 while shortcuts are needed there.