feat: prefer screen inversion via effects
On X11 we can invert a screen via RandR. But on Wayland we currently only do
this via effects. Prefer the inversion via effects on RandR too. This
simplifies our code and makes X11 and Wayland behave more similarly. Also the
effect is not that important, so a question might be if we should remove the
RandR code not all together.