feat: specify RandR inversion screen via root window only
When screen inversion via Xlib was added in ab86f0e8 the author tried to
specify the screen we apply the inversion to by the active client. This makes
sense if there are more than one screen but this nowadays is usually not the
case anymore because we do multi display via RandR CRTCs.

Additionally this is only fallback code so we ignore this case and directly get
the screen resources via the root window. This simplifies our code and removes
a dependency on the space.