feat(effect): disable zoom with XRender
We remove the XRender code in the zoom effect in order to simplify the code and
better integrate KWin changes which has removed XRender compositing altogether
some time ago. The zoom effect is this way not anymore available with XRender
compositing, but since XRender is considered to be a fallback anyway this is
acceptable.