refactor: handle first window mapping without signal
The wayland::window is set ready to be painted and added to the stacking order
only on first commit and when not being hidden. We can do this through a simple
check on the ready_for_painting variable in the normal commit handler instead
of listening for the windowShown signal.