refactor: add new Wayland windows to lists in a minimal way
Wayland windows don't need to be propagated to the XServer and they are
guaranteed to not yet be in the stacking order pre-stack. Also they are never
so called "tool windows", what is an X11-only concept. So we can simply add
them to the pre-stack of the stacking order and update without propagation of
new windows.