refactor: ignore raise on pending windows
If a window is just added it might not yet be in the stacking order pre-stack.
In this case before we added it to the pre-stack and continued raising it. But
this is not necessary. The window is raised later on anyway once it is shown.
Instead now we abort such a raise operation and by that can rely on another
invariant until the window is shown.