refactor: template spaces over base
The link between our win component and other modules is planned to become more
loose via class templates over the common base.

For that template the Wayland and X11 space over a generic Base. To achieve
this we also need to template the Wayland window over the space. All three
class definitions are moved to the header to allow that.