refactor: remove Wayland window Q_OBJECT macro
Remove the macro to prepare templating the class. We only needed it for
qobject_cast usage but this can be replaced with dynamic_cast. It is also
needed for registering the type in the integration tests but this registration
is unneeded in itself and can simply be removed.

The motivation for this removal is to afterwards make the Wayland window type a
class template.