feat(script): deprecate X11 window specific fullscreen signal
The clientFullScreenSet signal was only fired for X11 windows with the
additional information if the fullscreen change has been caused by a user
action or client logic.

It is assumed this is irrelevant information for most scripts and the signal is
deprecated what allows us to remove the additional signal in the x11::window
class.