fix(debug): read out scripting window properties
The Toplevel class has lost its Q_PROPERTY macros already long ago. This means
the debug console can not read out any properties from it like it used to.

For now improve this situation in parts by at least reading out the properties
from the scripting window in case this is defined. This won't work for windows
without control but is a start. In the long-run it would probably make sense to
create a separate debug window class that connects to any window like the
scripting window does to windows with control.