refactor: provide window ids as signal payloads
Replace the Toplevel arguments in signals of our space with an idea, that is
used to get the window from the space windows_map.

This is a necessary indirection as a future templated Toplevel can not be an
argument of a signal.

The two signals currentDesktopChanged and clientActivated are special cases as
we remove the payload altogether and instead ask receivers to get the active,
respctively move-resive window directly from the space instead.