refactor: store render window in shadow
Instead of storing the Toplevel, store a pointer to the render window. Since
the shadow is part of the render module that makes more sense on an
architectural level and we can still access the Toplevel properties through the
pointer in the render window. Additionally we don't need to switch the window
in the shadow anymore on copying over to the remnant.