refactor: reference render window in effect window
Nowadays the hierarchy of ownership between Toplevel, render and effect window
is clear and changes to that on remnant creation are unnecessary. Instead of
storing a pointer to the Toplevel instance in the effect window which then gets
changed on remnant creation we can also simply store the render window and
access the Toplevel instance through that.