ci: add Coverity Scan modeling file
A modeling file allows to reduce the number of false positives that Coverity
Scan produces. [1]

Such a file must be uploaded through the admin interface of our project on the
Coverity Scan Website [2] to take effect. We mirror it additionally here in the
source for keeping track of changes and allowing changes to it through our
usual contribution workflow.

For now we add an implementation for the QTest library function qVerify as this
function aborts a test on a false statement. Coverity Scan is not aware of this
because our build does not contain the Qt library implementation of the
function and so did report resource leaks etc. afterwards.

[1] https://scan.coverity.com/tune
[2] https://scan.coverity.com/projects/kwinft