refactor: store stacking order as value type in space
We don't actually need to store the stacking order via a smart pointer in
spaces nowadays anymore.

So store the stacking order as a value type instead of using a STL unique_ptr.
This reduces complexity and should improve memory locality.

The value type is only movable because of the composited QObject, but this was
the same with the smart pointer. Only functional change is that we don't reset
the object anymore early in win::clear_space. But this is at the very end of
the space lifetime anyway and runtime tests don't show any issues with that.