refactor(input): introduce idle class
Replace the client connection in the KIdleTime plugin with an internal
mechanism to query and inhibit user idle notifications.

The advantage besides the removal of the roundtrip through the Wayland
connection is that our new internal mechanism only requires to run a single
timer.

The listeners are grouped according to their idle durations and the remainders
of longer idle times are calculated at each timer timeout.