refactor(input): setup space from redirect ctor
Removing the separate call improves the API and invariants. We need to ensure
that the space is only called through the reference in redirect through from
code that is potentially called through creation of the redirect instance and
by that at creation of the space instance. But this feels natural, so we adapt
respective filter and other code in the input module.