refactor: own DPMS filter in input redirect
The DPMS filter depends like all other filters on the redirect object being
available.

While at first one thinks that the DPMS filter should be independent of the
space and the redirect and more naturally to put into the platform class it is
not the case in practice and hides the relationship.

Additionally this removes the last direct dependency on the redirect from the
platform except its redirect pointer.