refactor: remove timer based mouse polling
We now require XInput on X11 for getting information about the mouse/cursor
position so the timer is not in use anymore.

Now we get continously information about the mouse on Wayland as well as on
X11 so the extra calls to start and stop mouse polling are not necessary
anymore and can be removed as well.