fix: amend stub virtual keyboard D-Bus interface
In order to allow building the Plasma workspace together with KWinFT we
maintain a copy of the org.kde.kwin.VirtualKeyboard protocol that we do not
implement but install as usual.

Some properties were added to the protocol and Plasma workspace expects them
since [1] to exist, so we must add them to our copy of the protocol too.

[1] https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/2087