refactor: reduce X11 compositor loops
use a for ranges loop to reduce one iteration of algorithms when
checking for damaged windows. Also reset and fetch damaged windows
while we are already looping the render_stack. check for pending
repaints while looping the render stack and also after
getDamageRegionReply which might add new regions to repaint. this
removes the need for iterating space->windows yet again.