refactor: store windows in heterogeneous container
We move away from the virtual class interface as storage type. Instead we store
windows in a variant.

This increases in a way complexity in that we have to use the overload pattern
when accessing properties of windows in the heterogeneous container but we gain
stricter type safety from it and can decide on paths to take at compile time by
using constexpr if.