refactor: provide output to focus chain functions
Instead of getting the output in the focus chain functions, retrieve it
before that depending on the reference and provide it to the functions.

The usable focus candidate function then also only gets the output and not a
reference anymore. The caller of the function should check before any reference
window.