refactor(input): set deco on devices without arguments
Instead of providing the old and new deco, unset and set deco on devices while
keeping the deco unchanged in the focus struct. This simplifies the code.

We also check on the specific unset and set handler in the device now so we
don't have to implement stubs for them in devices without use for these.