refactor: remove decorationChanged signal
Instead update the cursor image directly in the deco setter and unsetter of the
pointer redirect. This in theory changes the logic to update the cursor image
one more time in between setting it from one deco to a second one, but it is
assumed this is not noticeable by the user.