fix: call space init function at the end of Wayland space construction
When initializing the initial virtual desktop is set. At this point the input
redirect and its objects must already exist what is being done at some point in
the ctor call. Because this was not the case the screens test crashed locally.

To ensure all all objects are available initialize the space at the end of the
ctor call.