fix: place on global default
Since 52131a93 the placement code was using a switch-statement. But in the
unknown and global default cases the fallthrough leads to returning directly
from within the no placement case. This is an unexpected behavior change.

Instead now handle the special cases before and afterwards handle remaining
cases in the switch statement.