test: run integration tests without dbus-run-session
The dbus-run-session wrapper made the tests freeze on developer machines. This
might be only a temporary regresssion in dbus or else but since we recommend to
wrap the whole ctest call anyway with dbus-run-session in our docs, we might as
well remove the wrapper on each explicit test execution for now.