fix: textX11Client leaked slot connection cause segfault
Lambda-function (slot) with captured local (stack-allocated) variable
is connected to the signal in testX11WindowId function.
The signal can be emitted randomly in testCaptionChanges function
and cause undefined behavior or segmentation fault
(due to access to unallocated captured variable).
Thus, better to disconnect the slot from the signal at the end of the test.