fix(input): check on pointer before updating
Pointer devices might go away so there is no pointer to access anymore. This
was triggered by some tests when a focus update was processed after the wlroots
test devices have already been removed.

This is only an intermediate fix. Long-term we should aim for disconnecting the
update function of a device class when there is no device of this class anymore.