refactor: initialize Night Color manager fully in ctor
Instead of waiting for the startup signal complete the initialization sequence
directly in the ctor. This should be in general possible nowadays since the
overall startup sequence of space, render backend, etc. has been improved. We
need to check on the compositor being created though on early gamma ramp
commits. This should be revisited in the future by either letting the
compositor own the Night Color manager or by further improving the startup
sequence.

The config loading must be done earlier now to create the settings.