feat: remove global default screen getter
Instead store it in the x11_data field of the base. The caching logic is put
into a separate free function in the base namespace. Later on we should check
if this caching logic makes sense at all or if we could just query the screen
once in the beginning explicitly.