feat: add rules KCM warning for low opacity
Once upon a time, in a moment of acute stupidity, I accidentally created a
window rule making all windows completely transparent. I meant to go for 0%
transparency when really the slider said 0% opacity. And once you've made the
very window to change this setting invisible, the only way to fix that mistake
is to log into a TTY or different DE and find the config file to edit manually.
Since I've since seen several KRedditors fall into the same trap:

Show a warning when active or inactive opacity is set to ~~<= 15%~~ < 25%.