refactor: provide internal X11 NET classes
Incorporate the NET parent and associated helper classes for internal
consumption instead of relying on KWindowSystem. We only need to include a
subset of funcitonality required for our window managing needs and few client
uses.

On the other side we continue to run the KWindowSystem plugin though in the
Wayland session for integration with Plasma.