feat: remove X11 startup notification support
Our internal support in the win module was only used for opening a window on a
specific desktop. Startup "feedback" as in showing an animated icon next to the
cursor is handled separately through the startup feedback effect, which uses
KWindowSystem for that.

If at some point we want the effect to be independent of KWindowSystem we would
need to add back internal support for startup notifactions, but this could then
be a reduced implementation only providing the bare minimum to forward the
feedback information to the effect.

Note that it is questionable if the previous startup notification backend
ported over from KWindowSystem was functioning correctly. So if support needs
to be readded this must be considered when looking at the code here removed.