refactor(input): interact with shortcuts through input interfaces
Instead of calling directly into KGlobalAccel use an interface in the input
module. In the future this should allow us to have KGlobalAccel as an
optional dependency only for our core infrastructure.

There should only be a single functional difference in that we ignore legacy
shortcuts for the Night Color manager now in order to not add additional API
for that in the input module.