fix(xwl): reset NET atoms on Xwayland shutdown
The atom value hash needs to be reset when the Xwayland connection gets lost.
This is due to the hash holding all atoms for a certain connection pointer,
but on restarting Xwayland that connection might have the same memory location
while having different atoms.

This is just a quick fix. Long-term we should look into making the atom hash a
non-global instance.