test: rewrite unit tests with Catch2
Like for our integration tests rewrite our unit tests with Catch2 instead of Qt
Test. We do this for all tests except the blend picture test which uses XRender
and is problematic because of the required X11 connection. Since the XRender
render backend has recently been removed we ignore the test.