test: generate operation mode variants
Many integration tests were run with Qt Test with and without Xwayland active.
Run these two variants again via the GENERATE macro of Catch2. We do this for
most tests that it was done for in the past and some additional ones where it
seems useful.

In case of new test variants not all of them pass at the moment for the
Xwayland variant. We abort these variants for now and mark them for later
inspection.