fix: set viewport depending on render target
When we render to the intermediate framebuffer texture the viewport must be set
first to the "magic" dimensions to capture the output area fromt the whole
rendering scene and in the second rendering step we need the simple viewport
onto the texture.

We also switch back the viewport y-position to a calculation that was done in
the past and proved to be functional.