refactor: separate render options into own class
Split up base::options further by adding a separate render::options class.

The base kcfg file is removed. It is also not installed anymore. We assume no
other project is accessing it at the moment. Otherwise we will need to add a
unified version of it back again in the future just for installing.