feat: remove compositing status of tabbox switcher item
This was added in 6a0f5016, but the patch actually making use of it [1] never
was merged to Plasma. The property never got used by any officially shipped
window switcher as a quick grep revealed.

In light of that and that the property would be useless on Wayland, we remove
it again. This removes a dependency from the win to the render module.

[1] https://invent.kde.org/plasma/kdeplasma-addons/-/merge_requests/69