refactor: grab X keyboard only on X11
Move the grab function to the input::x11 platform. The calls are only being
compiled when building for X11, not whith Xwayland anymore.

This means we omit the calls completely on Xwayland which is different to
before. But since we have control over grabs on Wayland anyway, that should be
the same.