fix(effect): make magiclamp look good with floating panels
The magic lamp literally "squashes" the window through the window icon
in the task manager.

It's assumed that there's nothing below the panel, so the magic lamp
doesn't perform any clipping.

With floating panels, it's not the case. So let's clamp the x or the y
coordinates when the window moves horizontally or vertically,
respectively, in order to ensure that the window is not visible in the
gap between the floating panel and the screen edge.

BUG: 361121
BUG: 466177