fix(effect): also do things when changing to active in overview
We were binding to the statusChanged signal, but then only doing
anything if the status had changed to "Activating". This did not work
for when the effect was activated without a gesture, because then its
status would change straight to "Activated", bypassing the in-between
state of "Activating."

This fixes clearing the search text on activation without a gesture.