fix: specify base config file name
The config file name must be set to "kwinrc", otherwise it will use the app
name, which is either "kwin_x11rc" or "kwin_waylandrc". For integration tests
we use an empty string so it acutally picks the test name.

In theory we could also hardcode the "kwinrc" string in the base config. But
for now let the consumer specify the name until we have defined more clearly
how the configs of different consumers should or should not interact with each
other.