fix: adapt offscreenquickview to Qt6 changes
Qt6 documentation states:

"Note: When using this window in combination with QQuickRenderControl,
the result of this function is an empty image, unless the software
backend is in use. "

We support a mode of using openGL rendering but then exporting as pixmap
images for Aurorae. In this case we should simply grab data from our
FBO.

BUG: 471041