fix: replay move window action if window isn't movable
Otherwise, when using Alt as modifier for moving windows, the tabbox window
cannot be clicked because Alt is held down during Alt+Tab, which would try
to initiate a window move but the window isn't actually movable, so ends up
doing nothing.

It also allows to interact with Plasma popups, which are also immovable,
while the modified key is pressed.