fix: correct highlight when hovering the preview
The preview item implements hover event filters, which by default
are accepted and not propagated to their parent, breaking the
hightlight on hover for the GridView.

We can simply "ignore()" those events so they are propagated,
but still fully functional.

Preemptively, do the same also for mouse events, which seem to be
already propagated now, but shouldn't according to the QQuickItem
documentation.

CCBUG: 473011