refactor: port screenedges kcm to KSvg
Removes the dependency on Plasma Framework from the KCMs.

This won't use the current Plasma theme but always the default one.
However, the monitor graphic hasn't really been used anywhere else
anymore (it used to be in e.g. the wallpaper settings in Plasma 4),
so I think this is alright.