refactor: store render targets in render struct
Instead of using a static stack, reference the render target stack in the
render struct that is moved through the painting process. This better
encapsulates our types.

To achieve this we introduce a new abstract render::framebuffer type. This
will also help us in the future with opening up to other rendering engines.