refactor: remove virtual desktop functors
The functors to get a VD in a specific direction are a neat idea that in some
cases can reduce code duplication. But at the moment there is just too much
noise with different methods of doing the same.

Another issue is that they hinder a future templating of the virtual desktop
manager class, or at least make it less attractive.

For now we remove the functor concept and just directly call into the
respective functions of the manager. We copy stubs of the functors over to the
virtual desktop integration test where these are still useful.