style: rename virtual desktops to subspaces
The name "virtual desktop" is longish. And it is too close to the "desktop" as
a window being such. In the end we also want this technology to be not only
about desktop usage.

In the future we may generalize the concept of subspaces to more than two
dimensions, then becoming indeed affine hyperplanes. Insofar the mathematical
terminology is approrpiate.