feat: store keyboard layouts independently of session manager
The session manager is an X11 concept. We should not abuse it for saving and
loading keyboard layouts. Instead read layouts from config directly in the
ctor of a layout policy and write them in the dtor.

For that we need to ensure that layouts are available at this point in time
including when a window was closed before. This is resaonable as this improves
our feature set now not only remembering layouts for windows that have exist
at the end of a session but also for windows being opened and later closed
again in general.