refactor: move xcb cursors cache to space
This removes an X11 dependency in the generic input::cursor code. It has the
downside that we have to add a signal connection to clear the cache on cursor
theme changes. For the moment we accept this, but in the future we should
reevaluate the relation between the cursor classes and the input module.