test: add Minico - the mini Wayland compositor
Add a minimum viable product Wayland compositor that is meant to showcase a
minimal configuration to get started with KWinFT libraries, similar to tinywl
for wlroots.

This compositor is currently functional in that a session spawns and clients
can connect by specifying the Wayland socket. Robustness and correctness of
operation must be further improved though in the future.

Another goal is to add additional features behind compile flags one-by-one
allowing to activate more features at compile time.