refactor: own QML engine in space
The QML engine is not only used for script components at the moment but also
for the on screen display, tabbox, rendered outline and QML-based effects.

While the last one is also partly depending on script components, the others
are not. Therefore move the QML engine out of the script platform and instead
own it in spaces, where we at least need to have it due to the on screen
display and tabbox.