feat(x11): remove alternative wm dialog
It's a neat idea to show a dialog to select a different WM when we are crashing
too often. But In practice only few people have installed more than one WM and
we shouldn't crash in the first place.

Instead of showing a dialog just exit directly so we don't retry ad infinitum
if there is really a fatal condition.