feat(x11): add wm selection owner type
Move the definition from the main_x11 source file to the library so taht it can
be used by other consumers too.

We use this opportunity to introduce a new base backend x11 library, since this
type is not useful with Xwayland. We'll see if this makes sense longterm but
for now let's try it out.