ci: use advanced artifact dl
Action [1] provides a more feature-rich API to download artifacts. This way we
can specify a repo and branch to download artifacts from. We need this in order
to download the artifacts produced most recently by the upstream Compositor
Modules. It is also helpful for running actions on WIP branches dependent on
another WIP branche in the upstream repo.

[1] https://github.com/dawidd6/action-download-artifact